// lex/types/com/atproto/repo/applyWrites.ts
export type QueryParams = globalThis.Record<PropertyKey, never>;

export interface InputSchema {
  /** The handle or DID of the repo (aka, current account). */
  repo: string;
  /** Can be set to 'false' to skip Lexicon schema validation of record data across all operations, 'true' to require it, or leave unset to validate only for known Lexicons. */
  validate?: boolean;
  writes: ($Typed<Create> | $Typed<Update> | $Typed<Delete>)[];
  /** If provided, the entire operation will fail if the current repo commit CID does not match this value. Used to prevent conflicting repo mutations. */
  swapCommit?: string;
}

export interface OutputSchema {
  commit?: ComAtprotoRepoDefs.CommitMeta;
  results?:
    ($Typed<CreateResult> | $Typed<UpdateResult> | $Typed<DeleteResult>)[];
}

export interface CallOptions {
  signal?: AbortSignal;
  headers?: HeadersMap;
  qp?: QueryParams;
  encoding?: "application/json";
}

export interface Response {
  success: boolean;
  headers: HeadersMap;
  data: OutputSchema;
}

export class InvalidSwapError extends XRPCError {
  constructor(src: XRPCError) {
    super(src.status, src.error, src.message, src.headers, { cause: src });
  }
}

export function toKnownErr(e: unknown): unknown {
  if (e instanceof XRPCError) {
    if (e.error === "InvalidSwap") return new InvalidSwapError(e);
  }

  return e;
}

/** Operation which creates a new record. */
export interface Create {
  $type?: "com.atproto.repo.applyWrites#create";
  collection: string;
  /** NOTE: maxLength is redundant with record-key format. Keeping it temporarily to ensure backwards compatibility. */
  rkey?: string;
  value: { [_ in string]: unknown };
}

const hashCreate = "create";

export function isCreate<V>(v: V): v is Create & V {
  return is$typed(v, id, hashCreate);
}

export function validateCreate<V>(v: V): ValidationResult<Create & V> {
  return validate<Create & V>(v, id, hashCreate);
}

/** Operation which updates an existing record. */
export interface Update {
  $type?: "com.atproto.repo.applyWrites#update";
  collection: string;
  rkey: string;
  value: { [_ in string]: unknown };
}

const hashUpdate = "update";

export function isUpdate<V>(v: V): v is Update & V {
  return is$typed(v, id, hashUpdate);
}

export function validateUpdate<V>(v: V): ValidationResult<Update & V> {
  return validate<Update & V>(v, id, hashUpdate);
}

/** Operation which deletes an existing record. */
export interface Delete {
  $type?: "com.atproto.repo.applyWrites#delete";
  collection: string;
  rkey: string;
}

const hashDelete = "delete";

export function isDelete<V>(v: V): v is Delete & V {
  return is$typed(v, id, hashDelete);
}

export function validateDelete<V>(v: V): ValidationResult<Delete & V> {
  return validate<Delete & V>(v, id, hashDelete);
}

export interface CreateResult {
  $type?: "com.atproto.repo.applyWrites#createResult";
  uri: string;
  cid: string;
  validationStatus?:
    | "valid"
    | "unknown"
    | (string & globalThis.Record<PropertyKey, never>);
}

const hashCreateResult = "createResult";

export function isCreateResult<V>(v: V): v is CreateResult & V {
  return is$typed(v, id, hashCreateResult);
}

export function validateCreateResult<V>(
  v: V,
): ValidationResult<CreateResult & V> {
  return validate<CreateResult & V>(v, id, hashCreateResult);
}

export interface UpdateResult {
  $type?: "com.atproto.repo.applyWrites#updateResult";
  uri: string;
  cid: string;
  validationStatus?:
    | "valid"
    | "unknown"
    | (string & globalThis.Record<PropertyKey, never>);
}

const hashUpdateResult = "updateResult";

export function isUpdateResult<V>(v: V): v is UpdateResult & V {
  return is$typed(v, id, hashUpdateResult);
}

export function validateUpdateResult<V>(
  v: V,
): ValidationResult<UpdateResult & V> {
  return validate<UpdateResult & V>(v, id, hashUpdateResult);
}

export interface DeleteResult {
  $type?: "com.atproto.repo.applyWrites#deleteResult";
}